remove trailing zeros

+++ cs/maths/conversions/molecular_chemistry.py
@@ -20,7 +20,7 @@ def molarity_to_normality(nfactor: int, moles: float, volume: float) -> float:
     >>> molarity_to_normality(4, 11.4, 5.7)
     8
     """
-    return round((float(moles / volume) * nfactor))
+    return round(moles / volume * nfactor)
 
 
 def moles_to_pressure(volume: float, moles: float, temperature: float) -> float:
@@ -40,7 +40,7 @@ def moles_to_pressure(volume: float, moles: float, temperature: float) -> float:
     >>> moles_to_pressure(8.2, 5, 200)
     10
     """
-    return round(float((moles * 0.0821 * temperature) / (volume)))
+    return round(moles * 0.0821 * temperature / volume)
 
 
 def moles_to_volume(pressure: float, moles: float, temperature: float) -> float:
@@ -60,7 +60,7 @@ def moles_to_volume(pressure: float, moles: float, temperature: float) -> float:
     >>> moles_to_volume(8.2, 5, 200)
     10
     """
-    return round(float((moles * 0.0821 * temperature) / (pressure)))
+    return round(moles * 0.0821 * temperature / pressure)
 
 
 def pressure_and_volume_to_temperature(
@@ -82,4 +82,4 @@ def pressure_and_volume_to_temperature(
     >>> pressure_and_volume_to_temperature(8.2, 5, 3)
     60
     """
-    return round(float((pressure * volume) / (0.0821 * moles)))
+    return round(pressure * volume / (0.0821 * moles))
